Added e10fuelproperties to the annual preagg script, weighting the properties with monthly VMT.
This fixes #4, where a data overflow issue could have resulted in incorrect PM emissions. This is because the e10fuelproperties table wasn't being used at annual preagg (since there were no annual [i.e., monthID 0] rows in it), and the model was falling back to values in the fuelformulation table. These values were never intended to be used, which caused the overflow. By adding this table to the preagg script, an annual average monthID 0 value is calculated for each fuel, and the e10fuelproperties table will be used when running in this mode.