fix for NR LF/AvgHP (#7) and NR diesel-only runs (#6)
For the NR LF/AvgHP: The Go calculator is implicitly splitting both emissions and activity by fuelSubTypeID when that output is requested by the user. The issue is that the LF and AvgHP activity values shouldn't be split, but should be duplicated. Since it is happening implicitly in the Go calculator, we decided to simply recombine the LF and AvgHP values as soon as the Go calculator results are loaded by the Java, instead of adding a separate edge case in the Go.

For NR diesel-only runs: When the user exports the default fuels, it only includes fuel types selected in the runspec. So when these fuels are reimported, there are no rows about gasoline. Unlike other tables, the nrfuelsupply table completely overwrites the default table (instead of being combined with the default table), so at runtime, there is no gasoline information. When the Java code is setting up the NR Fortran inputs, it doesn't have gasoline information, so it supplies 0 RVP. When the Fortran starts, it performs a validity check on the inputs, including valid RVP values (regardless if they will be used or not), and it quits when it sees an RVP of 0. The solution is to provide the Fortran with the default fuelFormulationID 10 RVP and sulfur levels when no gasoline values can be found. A warning is emitted when this occurs, so the user is aware and can make changes if this is unintentional.