#16 updating fuel importer logic to only apply E200/E300 calculations for gasoline/gasohol fuel types
The fuelformulation E200 and E300 numbers are not used for E-85 (they use the e10fuelproperties table, instead). Additionally, the approximation of these numbers doesn't make sense for diesel, CNG, or electricity. Therefore, only perform this calculation for fuelSubTypeIDs 10, 11, 12, 13, 14, 15.

Additionally, warn the user if they import diesel parameters that MOVES will not use.