#11 when importing VMT via XML, import all VMT tables presented to the importer
When run via the GUI, the various tables are imported or skipped based on whether or not the GUI element has been made visible or is hidden. When creating a new database, the user selects radio buttons that control which elements are visible (with the default option as HPMS/Annual). When viewing a database that already exists, the GUI determines which elements are visible based on which tables have already been imported.

When run via XML importers, the same GUI logic occurs. If the database doesn't exist, the default option is selected and the "visibility" of each element is set (even though nothing is actually displayed to the user). If the database does exist, the "visibility" it set based on which tables are already present. If an element is "invisible", that table will be skipped.

The above causes an issue if you are trying to import one of the "non-default" configuration of tables--e.g., sourceTypeYearVMT. By default, it is "hidden", so the XML importer skips loading this table.

The solution is to signal to the importer when it is being run from XML. This signal is set when the importer manager is created in the fromXML().

Then, the VMT importer should only skip a table iff it is run from the GUI and the table is not visible. Otherwise, it should should import any table that is presented to it.

Moreover, the VMT importer should clear all tables when run from XML because we don't know ahead of time which tables will be presented to it. If tables aren't cleared first, incompatible data could be loaded into the same table.

Note that this entire issue of only importing the default tables displayed to the user is only an issue for the VMT importer because it is the only importer that shows or hides tables based on a radio button selection that the user makes in the GUI. All other importers always show all tables available on that tab at all times.

Additionally, a fix for issue where every XML importer session ends with the message "Import complete, but no files were specified."

This message is displayed for the entire session, but is actually reflective of only the last importer to be run. The last importer is typically the Generic importer, which is usually not used.

To make the message reflective of the entire session (i.e., only say no tables were imported if NO tables were imported), the "foundAnyTables" variable was promoted to a static variable on the BasicDataHandler class. So if any instance of the data handler set "foundAnyTables" to true, then it will be true at the end when the message is emitted.