configuration updates to improve cross-os compatibility (#25)
Linux is case-sensitive and uses forward slashes. Since Windows is case insensitive and can use either forward or backward slashes, this commit updates the configuration files to be more strict about case and to use forward slashes.

Note that instead of updating all of the MariaDB table names to be the correct case, there is a MariaDB configuration setting that allows MariaDB to be case-insensitive on Linux: lower_case_table_names=1 (see readme.md).

Also improved the instructions on how to compile the various components of MOVES, including the Nonroad FORTRAN.