Fix for increasing brakewear with more source types and/or later CYs (#37).
This issue is caused by an incorrect join the BaseRateGenerator.

For background, at project scale, for each process, the SourceTypePhysics utility runs and then BaseRateGenerator.generateSBWeightedEmissionRates runs. Then for each link, BaseRateGenerator.generateBaseRates runs the Go generator.
* The SourceTypePhysics functions create "opModeIDOffset"s, which are essentially model year-specific operating modes. The updateEmissionRateTables function duplicates rows in EmissionRate for each of the offset opmodes.
* BaseRateGenerator.generateSBWeightedEmissionRates essentially joins the EmissionRate table with the SourceBinDistribution table--this basically distributes an emission rate across source types in a source bin (note that source bins just have fuel type, reg class, and model year in them). This distribution is made based on activity in each source bin. After this step, we have access to SourceTypeID, ModelYearID, FuelTypeID, and RegClassID columns
* The issue is that the join between EmissionRate and SourceBinDistribution did not correctly join the offset opmodes. The original code joined on the left two characters of opModeIDOffset, which fails when there are more than 100 rows in the sourceUseTypePhysics table.
* Specifically, there is an OpModeIDOffset for each row in SourceUseTypePhysics. If there are more than 99 rows of SourceUseTypePhysics that apply to the run, there will be too many rows in SBWeightedEmissionRates, which in turn produces BaseRates that are too high.

The fix is to mathematically compare the offset of the opModeIDOffset, which is calculated by taking the `floor()` of the opModeIDOffset/100 and then multiplying again by 100 since all the offsets are multiples of 100. We don't actually need to multiply by 100 again for the comparison, but including that step clarifies the intention of our comparison.

This commit also fixes a confusing debug message emitted by the SourceTypePhysics utility.