fix hotelling input with 0 restricted access VMT (#41)
At county scale, if the optional input HotellingHoursPerDay is supplied, along with a road type distribution that puts 0 VMT on roadTypeIDs 2 and 4, MOVES will ignore the user input and calculate 0 hotelling hours.

MOVES uses long-haul combination truck VMT on restricted access road types as a surrogate for hotelling activity. It always calculates default hotelling activity first, and then scales this activity if the user supplies local hotelling data. This design allows users to only supply individual dimensions of hotelling activity (such as just the number of hours, or just adjust monthly differences) without having to supply the full amount of detail that MOVES uses in its calculations.

However, if the default hotelling calculations result in 0 hotelling hours, then scaling those results to match the user input fails.

This fix allows MOVES to use VMT on roadTypeIDs 3 and 5 (unrestricted access roads) to calculate default distributions when the user supplies data in HotellingHoursPerDay and there is no VMT on road types 2 and 4.