fix for Multiple Hotelling Years issue #42
The Total Activity Generator was calculating VMT once for each year (and clearing out the previous calculations). However, the master loop calls the TAG for each year for each process. Therefore, the TAG gets called for each year multiple times if there are multiple processes in the run. So for the first process, everything should be calculated correctly. But after that, the TAG will only have the final year's VMT, so anything VMT based will be incorrect for anything but the final year. A similar issue was identified for starts earlier and has already been corrected.

Instead of calculating VMT only if the context year is > than the previously calculated year, we should calculate VMT if they are !=. We also need to keep these results (shouldDeletePriorYears needs to be false so that when the VMT is recalculated, all results are retained).

Also needed to add a TRUNCATE vmtByMYRoadHourFraction statement to prevent duplicate key issues.