use BOMInputStream from Apache Commons IO to read files with BOM encoding (#39)
Java does not natively read UTF-8 BOM encoded files. When importing data in UTF-8 BOM encoded CSVs, MOVES would silently fail (that is, it wouldn't tell you that it failed because of the BOM, it would just fail to import the data. When loading data via the GUI, you'd see an error about missing a column; when loading data via command line, you'd get no error message at all).

The Apache Commons IO package has a BOMInputStream that silently discards the BOM if present. This object does everything an InputStream can do, so it's fairly drop-in compatible--we simply wrap the FileInputStream object into a BOMInputStream object, and all the following code just works.

The only complication was when trying to make a BufferedReader object, which does not take an InputStream an input. The solution was to wrap the BOMInputStream object into an InputStreamReader, which can be used as an input into a BufferedReader.

The BOMInputStream was used in all instances in which user input files might possibly be read. It was not used for internal file manipulation used by MOVES--for example, when copying files, jaring/unjaring files for creating bundles, etc., since MOVES will not encode files in UTF-8 BOM encoding.