MOVES4-ReleaseCandidate2 with movesdb20230515
The MOVES4 Release Candidate is made available for testing purposes and for modelers to become familiar with functional changes between MOVES3 and MOVES4 before MOVES4.0.0 is released. Emission results of MOVES4.0.0 may differ from the results of this release candidate. Results from this version may not be used in work for state implementation plans, conformity determinations, or for any other regulatory purpose.