fix for XPKNeededFuelSupply constraint error (#58)
If multiple counties are selected in the RunSpec, and those counties are in the same fuel region, the NeededFuelSupply query returns duplicate rows, which we don't need. A later sql command adds the XPKNeededFuelSupply index, and that command fails due to the duplicates.

A simple solution is to add the DISTINCT modifier to the query, to indicate that we only want distinct rows.
Even though this could return multiple rows for each set of keys included in XPKNeededFuelSupply (which includes everything in this table but defaultFuelFormulationID), we know that there is only ever going to be 1 defaultFuelFormulationID per fuelTypeID, so this is sufficient.