fix for Incorrect results in County Scale & Temporal Preagg without passenger car (#60)
In the temporal preagg scripts, sometimes a single source type is used when calculating weighting factors (weighting together the source types is a very complicated and time consuming calculation). The original code hardcoded passenger cars (source type 21) as the source type to use. To ensure this method works when 21s aren't included in the runs, the function that builds the execution database purposefully does not filter by source type for tables where this happens -- this is why this method works at Default Scale, because there is always data for 21s in the default database.

Unfortunately, at County Scale, users do not have to provide data for source types that are not included in the run; therefore, if passenger cars are missing, there are no weighting factors to use. For example, calculating an average market share at Annual Preagg will result in NULL marketShare values if passenger cars are missing from the run. Then, when the base rate calculator weights emission rates by the market share, the results will be 0 due to Go reading a NULL marketShare as a 0 market share.

The fix is to select a source type that has values to weight by. To minimize changes, we should always select passenger cars to use as the weighting factor if possible. If not, we will the most populous source type according to the default database, except never select motorcycles unless those are the only source types in the run, since motorcycles typically have the most different temporal distributions of all source types. Therefore, this order is: 21, 31, 32, 52, 61, 54, 62, 43, 53, 41, 42, 51, 11.

Since temporal pre-aggregation no longer requires passenger cars, we can remove the special case exceptions in InputDataManager.java. However, this means that default scale runs with temporal preaggregation without passenger cars will have very slightly different results than runs with passenger cars, because passenger cars will only be used to weight the temporal distributions when they are present in the run; otherwise, MOVES will use a more relevant source type for the run.