Enable the Zone tab for the Default Scale data manager (#63)
If a single county is selected in the RunSpec, ZoneRoadType should have a SHOAllocFactor of 1.0 for each road type, and Zone should have 1.0 for all of the "AllocFactor" columns. This will, in effect, allocate 100% of the VMT and VPOP (supplied through the Vehicle Type VMT and Source Type Population tabs, respectively) to the selected county.

If more than one county is selected in the RunSpec:

Supply VMT and VPOP inputs that reflect the combined VMT and VPOP for all selected counties.
Calculate each individual county's proportion of the combined VMT.
In ZoneRoadType, supply these values for each county's SHOAllocFactor on each road type. (The SHOAllocFactor column should sum to 4.0, because there are 4 road types in this table.)
In Zone, supply these values for each county's "AllocFactor" columns.
If state preaggregation is used, the VMT and VPOP inputs should reflect the total state VMT and VPOP, and the ZoneRoadType and Zone tables should be populated following the instructions above for multiple counties, as if each county in the state was selected individually in the RunSpec. At runtime, MOVES will use these allocation factors as weighting values when preaggregating other inputs, such as meteorology data. The output will be at the state level.

Note: Default scale input databases, while they can be created via the Create Input Database panel, are supplied in the RunSpec via the Advanced Features panel.